+++ source/tasks/Clean_Markdown_Of_Public_Commands.build.ps1
@@ -1,0 +1,98 @@
+<#
+    .SYNOPSIS
+        This is a build task that generates conceptual help.
+
+    .PARAMETER ProjectPath
+        The root path to the project. Defaults to $BuildRoot.
+
+    .PARAMETER OutputDirectory
+        The base directory of all output. Defaults to folder 'output' relative to
+        the $BuildRoot.
+
+    .PARAMETER BuiltModuleSubdirectory
+        Sub folder where you want to build the Module to (instead of $OutputDirectory/$ModuleName).
+        This is especially useful when you want to build DSC Resources, but you don't want the
+        `Get-DscResource` command to find several instances of the same DSC Resources because
+        of the overlapping $Env:PSmodulePath (`$buildRoot/output` for the built module and `$buildRoot/output/RequiredModules`).
+
+        In most cases I would recommend against setting $BuiltModuleSubdirectory.
+
+    .PARAMETER VersionedOutputDirectory
+        Whether the Module is built with its versioned Subdirectory, as you would see it on a System.
+        For instance, if VersionedOutputDirectory is $true, the built module's ModuleBase would be: `output/MyModuleName/2.0.1/`
+
+    .PARAMETER ProjectName
+        The project name. Defaults to the empty string.
+
+    .PARAMETER SourcePath
+        The path to the source folder name. Defaults to the empty string.
+
+    .PARAMETER BuildInfo
+        The build info object from ModuleBuilder. Defaults to an empty hashtable.
+
+    .NOTES
+        This is a build task that is primarily meant to be run by Invoke-Build but
+        wrapped by the Sampler project's build.ps1 (https://github.com/gaelcolas/Sampler).
+
+        There are also parameters that are intentionally added to the task, that is so
+        that other tasks that are run prior can change the values for the parameters
+        through for example environment variables.
+#>
+param
+(
+    [Parameter()]
+    [System.String]
+    $ProjectPath = (property ProjectPath $BuildRoot),
+
+    [Parameter()]
+    [System.String]
+    $OutputDirectory = (property OutputDirectory (Join-Path $BuildRoot 'output')),
+
+    [Parameter()]
+    [System.String]
+    $BuiltModuleSubdirectory = (property BuiltModuleSubdirectory ''),
+
+    [Parameter()]
+    [System.Management.Automation.SwitchParameter]
+    $VersionedOutputDirectory = (property VersionedOutputDirectory $true),
+
+    [Parameter()]
+    [System.String]
+    $ProjectName = (property ProjectName ''),
+
+    [Parameter()]
+    [System.String]
+    $SourcePath = (property SourcePath ''),
+
+    [Parameter()]
+    [System.Collections.Hashtable]
+    $BuildInfo = (property BuildInfo @{ })
+)
+
+# Synopsis: Clean the markdown documentation of the public commands.
+Task Clean_Markdown_Of_Public_Commands {
+    [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('DscResource.AnalyzerRules/Measure-ParameterBlockParameterAttribute', '', Justification='For boolean values when using (property $true $false) fails in conversion between string and boolean when environment variable is used if set as advanced parameter ([Parameter()])')]
+    param
+    (
+        [Parameter()]
+        [System.String]
+        $DocOutputFolder = (property DocOutputFolder 'WikiContent')
+    )
+
+    # Get the vales for task variables, see https://github.com/gaelcolas/Sampler#task-variables.
+    . Set-SamplerTaskVariable
+
+    "`tDocs output folder path = '$DocOutputFolder'"
+    ""
+
+    $DocOutputFolder = Get-SamplerAbsolutePath -Path $DocOutputFolder -RelativeTo $OutputDirectory
+
+    $markdownFiles = Get-ChildItem -Path "$DocOutputFolder/*.md" -Exclude '[Hh]ome.md'
+
+    Write-Build -Color 'Magenta' -Text 'Cleaning the command documentation.'
+
+    foreach ($markdownFile in $markdownFiles)
+    {
+        $markdownFile | Edit-CommandDocumentation
+    }
+}